#31 move resnet to own module
